Delay initialization until rundeck has bootstrapped
