+++ rundeck-execution-mode-timer/src/main/groovy/com/rundeck/plugin/EnableLaterExecutionsPluginGrailsPlugin.groovy
@@ -5,6 +5,7 @@ import grails.plugins.*
 
 class EnableLaterExecutionsPluginGrailsPlugin extends Plugin {
 
+    def loadAfter = ['rundeckapp']
     // the version or versions of Grails the plugin is designed for
     def grailsVersion = "3.3.10 > *"
     // resources that are excluded from plugin packaging
